Remove `LoginViewModel`
I ran across it again recently and realized that it is just an unnecessary level of indirection. So I removed it, moving all of its necessary stored properties onto the modifier itself and removing the properties that weren't needed.
